H3: move some common example code into a shared module
