+++ quinn-h3/examples/shared/mod.rs
@@ -1,0 +1,84 @@
+use std::fmt;
+
+use failure::{bail, Error, Fail, ResultExt};
+use quinn_proto::crypto::rustls::{Certificate, CertificateChain, PrivateKey};
+use slog::{info, o, Drain, Logger};
+use std::{fs, io, path::PathBuf};
+
+pub type Result<T> = std::result::Result<T, Error>;
+
+pub struct PrettyErr<'a>(&'a dyn Fail);
+impl<'a> fmt::Display for PrettyErr<'a> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        fmt::Display::fmt(&self.0, f)?;
+        let mut x: &dyn Fail = self.0;
+        while let Some(cause) = x.cause() {
+            f.write_str(": ")?;
+            fmt::Display::fmt(&cause, f)?;
+            x = cause;
+        }
+        Ok(())
+    }
+}
+
+pub trait ErrorExt {
+    fn pretty(&self) -> PrettyErr<'_>;
+}
+
+impl ErrorExt for Error {
+    fn pretty(&self) -> PrettyErr<'_> {
+        PrettyErr(self.as_fail())
+    }
+}
+
+pub fn build_certs(
+    log: Logger,
+    key: &Option<PathBuf>,
+    cert: &Option<PathBuf>,
+) -> Result<(CertificateChain, Certificate, PrivateKey)> {
+    if let (Some(ref key_path), Some(ref cert_path)) = (key, cert) {
+        let key = fs::read(key_path).context("failed to read private key")?;
+        let key = quinn::PrivateKey::from_der(&key)?;
+        let cert_chain = fs::read(cert_path).context("failed to read certificate chain")?;
+        let cert = quinn::Certificate::from_der(&cert_chain)?;
+        let cert_chain = quinn::CertificateChain::from_certs(vec![cert.clone()]);
+        Ok((cert_chain, cert, key))
+    } else {
+        let dirs = directories::ProjectDirs::from("org", "quinn", "quinn-examples").unwrap();
+        let path = dirs.data_local_dir();
+        let cert_path = path.join("cert.der");
+        let key_path = path.join("key.der");
+        let (cert, key) = match fs::read(&cert_path).and_then(|x| Ok((x, fs::read(&key_path)?))) {
+            Ok(x) => x,
+            Err(ref e) if e.kind() == io::ErrorKind::NotFound => {
+                info!(log, "generating self-signed certificate");
+                let cert = rcgen::generate_simple_self_signed(vec!["localhost".into()]).unwrap();
+                let key = cert.serialize_private_key_der();
+                let cert = cert.serialize_der().unwrap();
+                fs::create_dir_all(&path).context("failed to create certificate directory")?;
+                fs::write(&cert_path, &cert).context("failed to write certificate")?;
+                fs::write(&key_path, &key).context("failed to write private key")?;
+                (cert, key)
+            }
+            Err(e) => {
+                bail!("failed to read certificate: {}", e);
+            }
+        };
+        let key = quinn::PrivateKey::from_der(&key)?;
+        let cert = quinn::Certificate::from_der(&cert)?;
+        Ok((
+            quinn::CertificateChain::from_certs(vec![cert.clone()]),
+            cert,
+            key,
+        ))
+    }
+}
+
+pub fn logger(label: String) -> Logger {
+    let decorator = slog_term::PlainSyncDecorator::new(std::io::stderr());
+    let drain = slog_term::FullFormat::new(decorator)
+        .use_original_order()
+        .build()
+        .fuse();
+    Logger::root(drain, o!("h3" => label))
+}
